+++ js/reaction.js
@@ -18,14 +18,13 @@ goog.module('ord.reaction');
 goog.module.declareLegacyNamespace();
 exports = {
   commit,
-  dirty,
   downloadReaction,
   freeze,
   initFromDataset,
   initFromReactionId,
-  listen,
   setupObserver,
   toggleAutosave,
+  undoSlowly,
   unloadReaction,
   updateSidebar,
   validateReaction,
@@ -48,7 +47,6 @@ goog.require('ord.temperature');
 goog.require('ord.uploads');
 goog.require('ord.utils');
 goog.require('ord.workups');
-goog.require('proto.ord.Dataset');
 goog.require('proto.ord.Reaction');
 
 // Remember the dataset and reaction we are editing.
@@ -69,13 +67,12 @@ exports.session = session;
  */
 function init(reaction) {
   // Initialize all the template popup menus.
-  $('.selector').each((index, node) => ord.utils.initSelector($(node)));
-  $('.optional_bool')
-      .each((index, node) => ord.utils.initOptionalBool($(node)));
+  $('.selector').each((index, node) => initSelector($(node)));
+  $('.optional_bool').each((index, node) => initOptionalBool($(node)));
   // Enable all the editable text fields.
   $('.edittext').attr('contentEditable', 'true');
   // Initialize all the validators.
-  $('.validate').each((index, node) => ord.utils.initValidateNode($(node)));
+  $('.validate').each((index, node) => initValidateNode($(node)));
   // Initialize validation handlers that don't go in "add" methods.
   initValidateHandlers();
   // Initailize tooltips.
@@ -84,14 +81,14 @@ function init(reaction) {
   // (see github.com/twbs/bootstrap/issues/22610)
   Popper.Defaults.modifiers.computeStyle.gpuAcceleration = false;
   // Show "save" on modifications.
-  listen('body');
+  ord.utils.listen('body');
   // Load Ketcher content into an element with attribute role="application".
   document.getElementById('ketcher-iframe').contentWindow.ketcher.initKetcher();
   // Initialize the UI with the Reaction.
   loadReaction(reaction);
-  clean();
+  ord.utils.clean();
   // Initialize the collaped/uncollapsed state of the fieldset groups.
-  $('.collapse').each((index, node) => ord.utils.initCollapse($(node)));
+  $('.collapse').each((index, node) => initCollapse($(node)));
   // Trigger reaction-level validation.
   validateReaction();
   // Initialize autosave being on.
@@ -109,7 +106,7 @@ async function initFromDataset(fileName, index) {
   session.fileName = fileName;
   session.index = index;
   // Fetch the Dataset containing the Reaction proto.
-  session.dataset = await getDataset(fileName);
+  session.dataset = await ord.utils.getDataset(fileName);
   const reaction = session.dataset.getReactionsList()[index];
   init(reaction);
 }
@@ -119,7 +116,7 @@ async function initFromDataset(fileName, index) {
  * @param {string} reactionId
  */
 async function initFromReactionId(reactionId) {
-  const reaction = await getReactionById(reactionId);
+  const reaction = await ord.utils.getReactionById(reactionId);
   // NOTE(kearnes): Without this next line, `reaction` will be
   // partial/incomplete, and I have no idea why.
   console.log(reaction.toObject());
@@ -128,42 +125,10 @@ async function initFromReactionId(reactionId) {
 }
 
 /**
- * Fetches a reaction as a serialized Reaction proto.
- * @param {string} reactionId The ID of the Reaction to fetch.
- * @return {!Promise<!Uint8Array>}
- */
-function getReactionById(reactionId) {
-  return new Promise(resolve => {
-    const xhr = new XMLHttpRequest();
-    xhr.open('GET', '/reaction/id/' + reactionId + '/proto');
-    xhr.responseType = 'arraybuffer';
-    xhr.onload = function() {
-      const bytes = new Uint8Array(xhr.response);
-      const reaction = proto.ord.Reaction.deserializeBinary(bytes);
-      resolve(reaction);
-    };
-    xhr.send();
-  });
-}
-
-/**
  * Sets the `ready` value to true.
  */
 function ready() {
   $('body').attr('ready', true);
-}
-
-/**
- * Adds a change handler to the given node that shows the 'save' button when
- * the node text is edited.
- * @param {!Node} node
- */
-function listen(node) {
-  ord.utils.addChangeHandler($(node), dirty);
-  $('.edittext', node).on('focus', event => ord.utils.selectText(event.target));
-  $('.floattext', node).on('blur', event => ord.utils.checkFloat(event.target));
-  $('.integertext', node)
-      .on('blur', event => ord.utils.checkInteger(event.target));
 }
 
 /**
@@ -185,37 +150,20 @@ function clickSave() {
  */
 function toggleAutosave() {
   // We keep track of timers by holding references, only if they're active.
-  const autosave = $('#toggle_autosave');
   if (!session.timers.short) {
     // Enable a simple timer that saves periodically.
     session.timers.short =
         setInterval(clickSave, 1000 * 15);  // Save after 15 seconds
-    autosave.text('autosave: on');
-    autosave.css('backgroundColor', 'lightgreen');
+    $('#toggle_autosave').text('autosave: on');
+    $('#toggle_autosave').css('backgroundColor', 'lightgreen');
   } else {
     // Stop the interval timer itself, then remove reference in order to
     // properly later detect that it's stopped.
     clearInterval(session.timers.short);
     session.timers.short = null;
-    autosave.text('autosave: off');
-    autosave.css('backgroundColor', 'pink');
+    $('#toggle_autosave').text('autosave: off');
+    $('#toggle_autosave').css('backgroundColor', 'pink');
   }
-}
-
-/**
- * Shows the 'save' button.
- */
-function dirty() {
-  $('#save').css('visibility', 'visible');
-}
-
-/**
- * Hides the 'save' button.
- */
-function clean() {
-  const save = $('#save');
-  save.css('visibility', 'hidden');
-  save.text('save');
 }
 
 /**
@@ -259,7 +207,7 @@ function commit() {
   const reaction = unloadReaction();
   const reactions = session.dataset.getReactionsList();
   reactions[session.index] = reaction;
-  ord.utils.putDataset(session.fileName, session.dataset);
+  putDataset(session.fileName, session.dataset);
   ord.uploads.putAll(session.fileName);
 }
 
@@ -284,25 +232,6 @@ function downloadReaction() {
 }
 
 /**
- * Downloads a dataset as a serialized Dataset proto.
- * @param {string} fileName The name of the dataset to fetch.
- * @return {!Promise<!Uint8Array>}
- */
-function getDataset(fileName) {
-  return new Promise(resolve => {
-    const xhr = new XMLHttpRequest();
-    xhr.open('GET', '/dataset/proto/read/' + fileName);
-    xhr.responseType = 'arraybuffer';
-    xhr.onload = function(event) {
-      const bytes = new Uint8Array(xhr.response);
-      const dataset = proto.ord.Dataset.deserializeBinary(bytes);
-      resolve(dataset);
-    };
-    xhr.send();
-  });
-}
-
-/**
  * Uploads a serialized Dataset proto.
  * @param {string} fileName The name of the new dataset.
  * @param {!proto.ord.Dataset} dataset
@@ -312,7 +241,7 @@ function putDataset(fileName, dataset) {
   const xhr = new XMLHttpRequest();
   xhr.open('POST', '/dataset/proto/write/' + fileName);
   const binary = dataset.serializeBinary();
-  xhr.onload = clean;
+  xhr.onload = ord.utils.clean;
   xhr.send(binary);
 }
 
@@ -416,6 +345,147 @@ function unloadReaction() {
   // Setter does nothing when passed an empty string.
   reaction.setReactionId($('#reaction_id').text());
   return reaction;
+}
+
+/**
+ * Reverses the hide() in the most recent invocation of removeSlowly().
+ * Removes the node's "undo" button. Does not trigger validation.
+ */
+function undoSlowly() {
+  $('.undoable').removeClass('undoable').show('slow');
+  $('.undo').not('#undo_template').hide('slow', function() {
+    $(this).remove();
+    updateSidebar();
+  });
+  ord.utils.dirty();
+}
+
+/**
+ * Marks the given node for possible future undo. Adds an "undo" button to do
+ * it. Deletes any preexisting undoable nodes and undo buttons.
+ * @param {!Node} node The DOM fragment to hide and re-show.
+ */
+function makeUndoable(node) {
+  $('.undoable').remove();
+  node.addClass('undoable');
+  $('.undo').not('#undo_template').remove();
+  const button = $('#undo_template').clone();
+  button.removeAttr('id');
+  node.after(button);
+  button.show('slow');
+}
+
+/**
+ * Toggles the visibility of all siblings of an element, or if a pattern is
+ * provided, toggles the visibility of all siblings of the nearest ancestor
+ * element matching the pattern.
+ * @param {!Node} node The element to toggle or use as the search root.
+ * @param {string} pattern The pattern to match for finding siblings to toggle.
+ */
+function toggleSlowly(node, pattern) {
+  node = $(node);
+  if (pattern) {
+    node = node.closest(pattern);
+  }
+  // 'collapsed' tag is used to hold previously collapsed siblings,
+  // and would be stored as node's next sibling;
+  // the following line checks whether a collapse has occured.
+  if (node.next('collapsed').length !== 0) {
+    // Need to uncollapse.
+    const collapsedNode = node.next('collapsed');
+    collapsedNode.toggle('slow', () => {
+      collapsedNode.children().unwrap();
+    });
+  } else {
+    // Need to collapse.
+    node.siblings().wrapAll('<collapsed>');
+    node.next('collapsed').toggle('slow');
+  }
+}
+
+/**
+ * Toggles the collapse of a section in the form.
+ * @param {string} button The element to toggle.
+ */
+function collapseToggle(button) {
+  $(button).toggleClass('fa-chevron-down fa-chevron-right');
+  toggleSlowly(button, 'legend');
+}
+
+/**
+ * Adds and populates a <select/> node according to its data-proto type
+ * declaration.
+ * @param {!Node} node A node containing a `data-proto` attribute.
+ */
+function initSelector(node) {
+  const protoName = node.attr('data-proto');
+  const protoEnum = ord.utils.nameToProto(protoName);
+  if (!protoEnum) {
+    console.log('missing require: "' + protoName + '"');
+  }
+  const types = Object.entries(protoEnum);
+  const select = $('<select>');
+  for (let i = 0; i < types.length; i++) {
+    const option = $('<option>').text(types[i][0]);
+    option.attr('value', types[i][1]);
+    if (types[i][0] === 'UNSPECIFIED') {
+      option.attr('selected', 'selected');
+    }
+    select.append(option);
+  }
+  node.append(select);
+}
+
+/**
+ * Sets up a three-way popup (true/false/unspecified).
+ * @param {!Node} node Target node for the new <select/> element.
+ */
+function initOptionalBool(node) {
+  const select = $('<select>');
+  const options = ['UNSPECIFIED', 'TRUE', 'FALSE'];
+  for (let i = 0; i < options.length; i++) {
+    const option = $('<option>').text(options[i]);
+    option.attr('value', options[i]);
+    if (options[i] === 'UNSPECIFIED') {
+      option.attr('selected', 'selected');
+    }
+    select.append(option);
+  }
+  node.append(select);
+}
+
+/**
+ * Sets up and initializes a collapse button by adding attributes into a div in
+ * reaction.html.
+ * @param {!Node} node Target node for the new button.
+ */
+function initCollapse(node) {
+  node.addClass('fa');
+  node.addClass('fa-chevron-down');
+  node.click(function() {
+    collapseToggle(this);
+  });
+  if (node.hasClass('starts_collapsed')) {
+    node.trigger('click');
+  }
+}
+
+/**
+ * Sets up a validator div (button, status indicator, error list, etc.) by
+ * inserting contents into a div in reaction.html.
+ * @param {!Node} oldNode Target node for the new validation elements.
+ */
+function initValidateNode(oldNode) {
+  let newNode = $('#validate_template').clone();
+  // Add attributes necessary for validation functions:
+  // Convert the placeholder onclick method into the button's onclick method.
+  $('.validate_button', newNode).attr('onclick', oldNode.attr('onclick'));
+  oldNode.removeAttr('onclick');
+  // Add an id to the button.
+  if (oldNode.attr('id')) {
+    $('.validate_button', newNode).attr('id', oldNode.attr('id') + '_button');
+  }
+  oldNode.append(newNode.children());
 }
 
 /**
@@ -569,6 +639,16 @@ function updateObserver() {
 }
 
 /**
+ * Scrolls the viewport to the selected input.
+ * @param {!Event} event
+ */
+function scrollToInput(event) {
+  const section = $(event.target).attr('input_name');
+  const target = $('.input[input_name=\'' + section + '\']');
+  target[0].scrollIntoView({behavior: 'smooth'});
+}
+
+/**
  * Updates the input entries in the sidebar.
  */
 function updateSidebar() {
@@ -584,7 +664,7 @@ function updateSidebar() {
     navNode.addClass('inputNavSection');
     navNode.attr('input_name', 'INPUT-' + name);
     $('#navInputs').append(navNode);
-    navNode.click(ord.utils.scrollToInput);
+    navNode.click(scrollToInput);
   });
   updateObserver();
 }
